EntityPropertyBinder.sortIndex is now a long, to allow the use of simple timestamps for default sorting.

+++ entity-module/src/main/java/com/foreach/across/modules/entity/bind/AbstractEntityPropertyBinder.java
@@ -55,7 +55,7 @@ abstract class AbstractEntityPropertyBinder implements EntityPropertyBinder
 
 	@Getter
 	@Setter
-	private int sortIndex;
+	private long sortIndex;
 
 	@Setter
 	private String binderPath;
