+++ entity-module/src/main/java/com/foreach/across/modules/entity/bind/DirectPropertyEntityPropertyBinder.java
@@ -38,12 +38,12 @@ final class DirectPropertyEntityPropertyBinder implements EntityPropertyBinder
 	private final EntityPropertyDescriptor descriptor;
 
 	@Override
-	public int getSortIndex() {
+	public long getSortIndex() {
 		return 0;
 	}
 
 	@Override
-	public void setSortIndex( int sortIndex ) {
+	public void setSortIndex( long sortIndex ) {
 
 	}
